python_advance/check_certificates/requirements.txt
cffi==1.14.0
cryptography==2.9
pycparser==2.20
pyOpenSSL==19.1.0
six==1.14.0
